can only quit when round not running, also push gamemodel after removal

+++ src/main/java/ch/uzh/ifi/hase/soprafs24/service/GameService.java
@@ -91,6 +91,7 @@ public class GameService {
     }
 
     public boolean removePlayerFromGame(Game game, long userId) {
+        if(game.isRoundRunning()) throw new ResponseStatusException(HttpStatus.CONFLICT, "Round is still running");
         game.removePlayer(userId);
         playerService.removePlayer(userId);
         gameRepository.save(game);
